pytest svm setup
