+++ setup.cfg
@@ -51,6 +51,7 @@ install_requires =
 [options.extras_require]
 test =
     pytest-astropy
+    #pytest-datafiles
 docs =
     sphinx-astropy
 dev =
