+++ tests/test_files/svm/test_img_data.csv
@@ -1,0 +1,13 @@
+numexp,rms_ra,rms_dec,nmatches,point,segment,gaia,det,wcs,cat,label,index
+20,25.89462642224672,23.52141153142447,28,1705,4285,32.0,1,0,2,0.0,hst_11099_04_wfc3_ir_total_ia0m04
+1,2.81405187141158,6.422889786759978,5,0,0,9.0,3,0,1,0.0,hst_12062_eh_wfc3_uvis_total_ibeveh
+4,13.8704220799427,12.1561334355269,64,1332,1493,129.0,3,0,1,1.0,hst_12109_01_wfc3_uvis_total_ibfn01
+7,9.126496172300612,14.767202388527,7,2185,1818,10.0,4,0,1,0.0,hst_8992_03_acs_wfc_total_j8cw03
+4,0.0,0.0,0,18,30,1.0,0,1,1,1.0,hst_9454_11_acs_hrc_total_j8ff11
+4,9.123720587868435,9.741267631304435,22,334,627,24.0,4,0,2,0.0,hst_9836_33_acs_wfc_total_j8rq33
+8,17.55323976980026,20.153275686666017,47,625,2381,349.0,2,0,5,0.0,hst_10183_03_acs_sbc_total_j8y503
+9,0.0,0.0,1,53,36,1.0,2,0,1,1.0,hst_13483_21_acs_sbc_total_jcdb21
+6,-1.0,-1.0,0,2,4,0.0,0,2,0,0.0,hst_8992_53_acs_hrc_total_j8cw53
+20,44.659013915663145,35.05632691764301,28,1974,0,32.0,1,0,2,1.0,hst_11099_04_wfc3_ir_total_ia0m04_f160w_all_stoc
+7,9.964327386570638,17.569769003001152,6,2149,5338,10.0,4,0,1,1.0,hst_8992_03_acs_wfc_total_j8cw03_f475w_clear2l_sub_stat
+4,19.71635474842033,25.3808403744063,21,336,1073,24.0,4,0,2,1.0,hst_9836_33_acs_wfc_total_j8rq33_clear1s_f250w_sub_stoc
