+++ tox.ini
@@ -41,7 +41,7 @@ deps =
     cov: coverage
     cov: pytest-cov
 
-    oldestdeps: numpy==1.21.*
+    #oldestdeps: numpy==1.21.*
     oldestdeps: astropy==4.0.3
 
     devdeps: git+https://github.com/astropy/astropy.git#egg=astropy
